Split remote workspace descriptor values
The SSH restore slice added standalone remote descriptor/configuration code to already-budgeted Swift files. Move those value types into their own source file and register it with the Xcode project so the file-length guard stays green without accepting new debt.

Constraint: Workflow guard enforces a Swift file length budget for Workspace.swift and SessionPersistence.swift.

Rejected: Raise the file-length budget | avoid accepting new debt for descriptor-only code

Confidence: high

Scope-risk: narrow

Tested: python3 scripts/swift_file_length_budget.py --budget .github/swift-file-length-budget.tsv

Tested: git diff --check

Not-tested: Local unit/UI tests by repository policy.
